feat: fix api paths, add api docs
