+++ app/components/Icon/index.js
@@ -14,7 +14,6 @@ import {
   faChevronDown,
   faGear,
   faSortDown,
-  faChevronLeft,
 } from '@fortawesome/free-solid-svg-icons';
 import { faFacebookF, faInstagram, faTwitter, faYoutube } from '@fortawesome/free-brands-svg-icons';
 const iconMap = {
@@ -34,7 +33,6 @@ const iconMap = {
   twitter: faTwitter,
   youtube: faYoutube,
   'sort-down': faSortDown,
-  'chevron-left': faChevronLeft,
 };
 
 export default function Icon({ type, ...props }) {
